Import of DURIST/Proc-ProcessTable-0.39 from CPAN.
gitpan-cpan-distribution: Proc-ProcessTable
gitpan-cpan-version:      0.39
gitpan-cpan-path:         DURIST/Proc-ProcessTable-0.39.tar.gz
gitpan-cpan-author:       DURIST
gitpan-cpan-maturity:     released

+++ os/aix.h
@@ -2,6 +2,7 @@
 /* Includes                                                                */
 /***************************************************************************/
 
+#include <stdlib.h>
 #include <procinfo.h>
 #include <unistd.h>
 #include <string.h>
@@ -68,7 +69,7 @@
 /* Globals                                                                 */
 /***************************************************************************/
 
-static long Sysmem;
+static unsigned long long Sysmem;
 static int  PageSize;
 static int  ProcessNumber;
